+++ internal/service/wiki/write_page.go
@@ -71,22 +71,5 @@ func validateWikiWritePageRequest(request *pb.WikiWritePageRequest) error {
 		return fmt.Errorf("empty Format")
 	}
 
-	commitDetails := request.GetCommitDetails()
-	if commitDetails == nil {
-		return fmt.Errorf("empty CommitDetails")
-	}
-
-	if len(commitDetails.GetName()) == 0 {
-		return fmt.Errorf("empty CommitDetails.Name")
-	}
-
-	if len(commitDetails.GetEmail()) == 0 {
-		return fmt.Errorf("empty CommitDetails.Email")
-	}
-
-	if len(commitDetails.GetMessage()) == 0 {
-		return fmt.Errorf("empty CommitDetails.Message")
-	}
-
-	return nil
+	return validateRequestCommitDetails(request)
 }