+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Implement WikiDeletePage RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/414
 - Implement WikiFindPage RPC
   https://gitlab.com/gitlab-org/gitaly/merge_requests/419
 - Update gitlab_git to b3ba3996e0bd329eaa574ff53c69673efaca6eef and set